rename role component name
