+++ src/web/pages/roles/DetailsPage.jsx
@@ -39,7 +39,7 @@ import withEntityContainer, {
   permissionsSubjectFilter,
 } from 'web/entity/withEntityContainer';
 import useTranslation from 'web/hooks/useTranslation';
-import RoleComponent from 'web/pages/roles/Component';
+import RoleComponent from 'web/pages/roles/RoleComponent';
 import RoleDetails from 'web/pages/roles/Details';
 import {
   selector as permissionsSelector,
